# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Observatory Observation Service
  version: 0.1.0
  description: Observation layer API for LLM agents.
servers:
  - url: http://localhost:52033
security:
  - bearerAuth: []
  - {}
paths:
  /health:
    get:
      security: []
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /v1/observe:
    post:
      summary: Observe a topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObserveRequest"
            example: { topic: Oracle Cloud Backup }
      responses:
        "200":
          description: Evidence-backed observation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ObserveResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/ObservationFailed" }
  /v1/observe-url:
    post:
      summary: Observe one URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
            example: { url: https://example.com }
      responses:
        "200":
          description: Extracted URL observation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ObserveUrlResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/ObservationFailed" }
  /v1/observation:
    get:
      summary: Recall observation history
      parameters:
        - name: topic
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Matching history
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecallResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid API key
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ObservationFailed:
      description: Search, fetch, or extraction failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok] }
    ObserveRequest:
      type: object
      required: [topic]
      properties:
        topic: { type: string, minLength: 1 }
    Source:
      type: object
      required: [title, url]
      properties:
        title: { type: string }
        url: { type: string, format: uri }
        fetched_at: { type: string, format: date-time }
        content_hash: { type: string, example: sha256:abc123 }
        extractor: { type: string, example: mozilla-readability }
        search_rank: { type: integer, minimum: 1 }
        snippet: { type: string }
    Evidence:
      type: object
      required: [source, claim]
      properties:
        source: { type: string, format: uri }
        claim: { type: string }
        sources: { type: array, items: { type: string, format: uri } }
        confidence: { type: number, minimum: 0, maximum: 1 }
    ObserveResponse:
      type: object
      required:
        [
          topic,
          query,
          summary,
          confidence,
          observed_at,
          cache_hit,
          source_count,
          sources,
          evidence,
        ]
      properties:
        topic: { type: string }
        query: { type: string }
        summary: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        observed_at: { type: string, format: date-time }
        cache_hit: { type: boolean }
        source_count: { type: integer, minimum: 0 }
        sources: { type: array, items: { $ref: "#/components/schemas/Source" } }
        evidence:
          { type: array, items: { $ref: "#/components/schemas/Evidence" } }
    ObserveUrlResponse:
      allOf:
        - $ref: "#/components/schemas/ObserveResponse"
        - type: object
          required: [title, content, source]
          properties:
            title: { type: string }
            content: { type: string }
            source: { $ref: "#/components/schemas/Source" }
    RecallResponse:
      type: object
      required: [history]
      properties:
        history:
          type: array
          items:
            type: object
            required: [summary, confidence, observed_at]
            properties:
              summary: { type: string }
              confidence: { type: number }
              observed_at: { type: string, format: date-time }
    Error:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }